backend/requirements.txt
asgiref==3.5.2
click==8.1.3
Django==4.1.2
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
npm==0.1.1
optional-django==0.1.0
PyYAML==6.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
Werkzeug==2.2.2
